Bamboo 10.x compatibility
