+++ src/main/java/org/jfrog/bamboo/builder/IvyDataHelper.java
@@ -4,6 +4,7 @@ import com.atlassian.bamboo.process.EnvironmentVariableAccessor;
 import com.atlassian.bamboo.task.CommonTaskContext;
 import com.atlassian.bamboo.task.TaskContext;
 import org.jfrog.bamboo.admin.ServerConfig;
+import org.jfrog.bamboo.admin.ServerConfigManager;
 import org.jfrog.bamboo.configuration.BuildParamsOverrideManager;
 import org.jfrog.bamboo.context.PackageManagersContext;
 import org.jfrog.build.extractor.clientConfiguration.ArtifactoryClientConfiguration;
@@ -15,8 +16,8 @@ import java.util.Map;
  */
 public class IvyDataHelper extends MavenAndIvyBuildInfoDataHelperBase {
 
-    public IvyDataHelper(BuildParamsOverrideManager buildParamsOverrideManager, CommonTaskContext context, PackageManagersContext buildContext, EnvironmentVariableAccessor envVarAccessor, String artifactoryPluginVersion, boolean aggregateBuildInfo) {
-        super(buildParamsOverrideManager, (TaskContext) context, buildContext, envVarAccessor, artifactoryPluginVersion, aggregateBuildInfo);
+    public IvyDataHelper(BuildParamsOverrideManager buildParamsOverrideManager, CommonTaskContext context, PackageManagersContext buildContext, EnvironmentVariableAccessor envVarAccessor, String artifactoryPluginVersion, boolean aggregateBuildInfo, ServerConfigManager serverConfigManager) {
+        super(buildParamsOverrideManager, (TaskContext) context, buildContext, envVarAccessor, artifactoryPluginVersion, aggregateBuildInfo, serverConfigManager);
     }
 
     @Override
