+++ src/main/java/org/jfrog/bamboo/task/ArtifactoryDeploymentUploadTask.java
@@ -6,6 +6,7 @@ import com.atlassian.bamboo.task.TaskException;
 import com.atlassian.bamboo.task.TaskResult;
 import com.atlassian.bamboo.task.TaskResultBuilder;
 import com.atlassian.bamboo.variable.CustomVariableContext;
+import com.atlassian.plugin.spring.scanner.annotation.imports.ComponentImport;
 import org.apache.commons.lang.StringUtils;
 import org.jetbrains.annotations.NotNull;
 import org.jfrog.bamboo.admin.ServerConfig;
@@ -20,6 +21,7 @@ import org.jfrog.bamboo.util.deployment.LegacyDeploymentUtils;
 import org.jfrog.build.extractor.clientConfiguration.ArtifactoryManagerBuilder;
 import org.jfrog.build.extractor.clientConfiguration.util.spec.SpecsHelper;
 
+import javax.inject.Inject;
 import java.io.File;
 import java.io.IOException;
 import java.util.HashMap;
@@ -32,9 +34,14 @@ import java.util.Map;
  */
 public class ArtifactoryDeploymentUploadTask extends ArtifactoryDeploymentTaskType {
 
+    @Inject
+    @ComponentImport
     private CustomVariableContext customVariableContext;
     private ServerConfig uploadServerConfig;
     private String fileSpec;
+
+    @Inject
+    private ServerConfigManager serverConfigManager;
 
     @Override
     protected void initTask(@NotNull CommonTaskContext context) throws TaskException {
@@ -45,7 +52,7 @@ public class ArtifactoryDeploymentUploadTask extends ArtifactoryDeploymentTaskTy
         uploadServerConfig = TaskUtils.getResolutionServerConfig(
                 deploymentUploadContext.getOverriddenUsername(runtimeContext, buildInfoLog, true),
                 deploymentUploadContext.getOverriddenPassword(runtimeContext, buildInfoLog, true),
-                ServerConfigManager.getInstance(), selectedServerConfig, new BuildParamsOverrideManager(customVariableContext));
+                serverConfigManager, selectedServerConfig, new BuildParamsOverrideManager(customVariableContext));
     }
 
     @NotNull
@@ -95,7 +102,6 @@ public class ArtifactoryDeploymentUploadTask extends ArtifactoryDeploymentTaskTy
      * Get configurations of the selected server in the task definition.
      */
     private ServerConfig getSelectedServerConfig(@NotNull CommonTaskContext deploymentTaskContext) {
-        ServerConfigManager serverConfigManager = ServerConfigManager.getInstance();
         String serverId = deploymentTaskContext.getConfigurationMap().get(ArtifactoryDeploymentUploadConfiguration.DEPLOYMENT_PREFIX + PackageManagersContext.SERVER_ID_PARAM);
         if (StringUtils.isBlank(serverId)) {
             // Compatibility with version 1.8.0
@@ -126,5 +132,9 @@ public class ArtifactoryDeploymentUploadTask extends ArtifactoryDeploymentTaskTy
 
     public void setCustomVariableContext(CustomVariableContext customVariableContext) {
         this.customVariableContext = customVariableContext;
+    }
+
+    public void setServerConfigManager(ServerConfigManager serverConfigManager) {
+        this.serverConfigManager = serverConfigManager;
     }
 }
