+++ src/main/java/org/jfrog/bamboo/task/ArtifactoryIvyTask.java
@@ -16,6 +16,7 @@ import org.apache.commons.lang.SystemUtils;
 import org.apache.tools.ant.types.Commandline;
 import org.jetbrains.annotations.NotNull;
 import org.jfrog.bamboo.admin.ServerConfig;
+import org.jfrog.bamboo.admin.ServerConfigManager;
 import org.jfrog.bamboo.builder.BuilderDependencyHelper;
 import org.jfrog.bamboo.builder.IvyDataHelper;
 import org.jfrog.bamboo.builder.MavenAndIvyBuildInfoDataHelperBase;
@@ -24,6 +25,7 @@ import org.jfrog.bamboo.util.PluginProperties;
 import org.jfrog.bamboo.util.TaskUtils;
 import org.jfrog.bamboo.util.Utils;
 
+import javax.inject.Inject;
 import java.io.File;
 import java.io.IOException;
 import java.util.Arrays;
@@ -41,26 +43,18 @@ public class ArtifactoryIvyTask extends BaseJavaBuildTask {
     public static final String TASK_NAME = "artifactoryIvyTask";
     public static final String EXECUTABLE_NAME = SystemUtils.IS_OS_WINDOWS ? "ant.bat" : "ant";
     private static final String IVY_KEY = "system.builder.ivy.";
-    private final EnvironmentVariableAccessor environmentVariableAccessor;
-    private final CapabilityContext capabilityContext;
-    private final BuilderDependencyHelper dependencyHelper;
+    private CapabilityContext capabilityContext;
+    private BuilderDependencyHelper dependencyHelper;
     private IvyBuildContext ivyBuildContext;
     private MavenAndIvyBuildInfoDataHelperBase ivyDataHelper;
     private String artifactoryPluginVersion;
-
-    public ArtifactoryIvyTask(final ProcessService processService,
-                              final EnvironmentVariableAccessor environmentVariableAccessor, final CapabilityContext capabilityContext,
-                              TestCollationService testCollationService) {
-        super(testCollationService, environmentVariableAccessor, processService);
-        this.environmentVariableAccessor = environmentVariableAccessor;
-        this.capabilityContext = capabilityContext;
-        dependencyHelper = new BuilderDependencyHelper("artifactoryIvyBuilder");
-        ContainerManager.autowireComponent(dependencyHelper);
-    }
+    @Inject
+    private ServerConfigManager serverConfigManager;
 
     @Override
     protected void initTask(@NotNull CommonTaskContext context) throws TaskException {
         super.initTask(context);
+        dependencyHelper = new BuilderDependencyHelper("artifactoryIvyBuilder");
         Map<String, String> combinedMap = Maps.newHashMap();
         combinedMap.putAll(context.getConfigurationMap());
         combinedMap.putAll(((TaskContext)context).getBuildContext().getBuildDefinition().getCustomConfiguration());
@@ -68,7 +62,7 @@ public class ArtifactoryIvyTask extends BaseJavaBuildTask {
         initEnvironmentVariables(ivyBuildContext);
         aggregateBuildInfo = ivyBuildContext.shouldAggregateBuildInfo(context, ivyBuildContext.getArtifactoryServerId());
         artifactoryPluginVersion = Utils.getPluginVersion(pluginAccessor);
-        ivyDataHelper = new IvyDataHelper(buildParamsOverrideManager, context, ivyBuildContext, environmentVariableAccessor, artifactoryPluginVersion, aggregateBuildInfo);
+        ivyDataHelper = new IvyDataHelper(buildParamsOverrideManager, context, ivyBuildContext, environmentVariableAccessor, artifactoryPluginVersion, aggregateBuildInfo, serverConfigManager);
     }
 
     @Override
@@ -176,5 +170,13 @@ public class ArtifactoryIvyTask extends BaseJavaBuildTask {
 
         return dependencyHelper.downloadDependenciesAndGetPath(bambooTmp, getPlanKey(customVariableContext), context,
                 PluginProperties.getPluginProperty(PluginProperties.IVY_DEPENDENCY_FILENAME_KEY));
+    }
+
+    public void setServerConfigManager(ServerConfigManager serverConfigManager) {
+        this.serverConfigManager = serverConfigManager;
+    }
+
+    public void setCapabilityContext(CapabilityContext capabilityContext) {
+        this.capabilityContext = capabilityContext;
     }
 }
